Track committer additions with git history
After https://github.com/NixOS/nixpkgs-committers/pull/31, people who
received commit access since a year ago were tracked with the date of
commit bit reception.

While this works, it's not necessary in the long term, because the git
history will keep track of when commit bits are given.

This change starts using the commit history when possible, which allows
avoiding the need for adding addition dates for new committers.