scripts/retire.sh: use commits endpoint rather than activity
The activity endpoint does not work with the merge queue, and counts
things like branch creations that we don’t really want to include.

This finds two‐parent merge commits with the correct author committed
and signed by the GitHub key that are reachable from the main branch;
I believe that PR merges are the only such commits, but any exceptions
should be very marginal.

Closes: #69